--- docs/auto_instrumentation/auto_instrumentation.rst
@@ -1,15 +1,0 @@
-OpenTelemetry Python Autoinstrumentation
-========================================
-
-.. automodule:: opentelemetry.auto_instrumentation
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-Submodules
-----------
-
-.. toctree::
-    :maxdepth: 1
-
-    instrumentor
