--- docs/auto_instrumentation/instrumentor.rst
@@ -1,7 +1,0 @@
-opentelemetry.auto_instrumentation.instrumentor package
-=======================================================
-
-.. automodule:: opentelemetry.auto_instrumentation.instrumentor
-    :members:
-    :undoc-members:
-    :show-inheritance:
