+++ docs/instrumentation/instrumentation.rst
@@ -1,0 +1,15 @@
+OpenTelemetry Python Instrumentation
+====================================
+
+.. automodule:: opentelemetry.instrumentation
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+    :maxdepth: 1
+
+    instrumentor
