+++ docs/instrumentation/instrumentor.rst
@@ -1,0 +1,7 @@
+opentelemetry.instrumentation.instrumentor package
+==================================================
+
+.. automodule:: opentelemetry.instrumentation.instrumentor
+    :members:
+    :undoc-members:
+    :show-inheritance:
