+++ ext/opentelemetry-ext-django/src/opentelemetry/ext/django/__init__.py
@@ -16,9 +16,9 @@ from logging import getLogger
 
 from django.conf import settings
 
-from opentelemetry.auto_instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.configuration import Configuration
 from opentelemetry.ext.django.middleware import _DjangoMiddleware
+from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 
 _logger = getLogger(__name__)
