+++ ext/opentelemetry-ext-flask/setup.cfg
@@ -42,7 +42,7 @@ packages=find_namespace:
 install_requires =
     flask ~= 1.0
     opentelemetry-ext-wsgi == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     opentelemetry-api == 0.9.dev0
 
 [options.extras_require]
