+++ ext/opentelemetry-ext-jinja2/setup.cfg
@@ -40,7 +40,7 @@ package_dir=
 packages=find_namespace:
 install_requires =
     opentelemetry-api == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     jinja2~=2.7
     wrapt >= 1.0.0, < 2.0.0
