+++ ext/opentelemetry-ext-pymongo/setup.cfg
@@ -41,7 +41,7 @@ package_dir=
 packages=find_namespace:
 install_requires =
     opentelemetry-api == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     pymongo ~= 3.1
 
 [options.extras_require]
