+++ ext/opentelemetry-ext-pymysql/setup.cfg
@@ -42,7 +42,7 @@ packages=find_namespace:
 install_requires =
     opentelemetry-api == 0.9.dev0
     opentelemetry-ext-dbapi == 0.9.dev0
-    opentelemetry-auto-instrumentation == 0.9.dev0
+    opentelemetry-instrumentation == 0.9.dev0
     PyMySQL ~=  0.9.3
 
 [options.extras_require]
