--- opentelemetry-auto-instrumentation/README.rst
@@ -1,19 +1,0 @@
-OpenTelemetry Auto Instrumentation
-==================================
-
-|pypi|
-
-.. |pypi| image:: https://badge.fury.io/py/opentelemetry-auto-instrumentation.svg
-   :target: https://pypi.org/project/opentelemetry-auto-instrumentation/
-
-Installation
-------------
-
-::
-
-    pip install opentelemetry-auto-instrumentation
-
-References
-----------
-
-* `OpenTelemetry Project <https://opentelemetry.io/>`_
