+++ tox.ini
@@ -12,9 +12,9 @@ envlist =
     py3{4,5,6,7,8}-test-sdk
     pypy3-test-sdk
 
-    ; opentelemetry-auto-instrumentation
-    py3{4,5,6,7,8}-test-auto-instrumentation
-    pypy3-test-auto-instrumentation
+    ; opentelemetry-instrumentation
+    py3{5,6,7,8}-test-instrumentation
+    pypy3-test-instrumentation
 
     ; opentelemetry-example-app
     py3{4,5,6,7,8}-test-example-app
@@ -146,7 +146,7 @@ setenv =
 changedir =
   test-api: opentelemetry-api/tests
   test-sdk: opentelemetry-sdk/tests
-  test-auto-instrumentation: opentelemetry-auto-instrumentation/tests
+  test-instrumentation: opentelemetry-instrumentation/tests
   test-ext-grpc: ext/opentelemetry-ext-grpc/tests
   test-ext-aiohttp-client: ext/opentelemetry-ext-aiohttp-client/tests
   test-ext-requests: ext/opentelemetry-ext-requests/tests
@@ -182,20 +182,19 @@ commands_pre =
 ; cases but it saves a lot of boilerplate in this file.
   test: pip install {toxinidir}/opentelemetry-api {toxinidir}/opentelemetry-sdk {toxinidir}/tests/util
 
-  test-auto-instrumentation: pip install {toxinidir}/opentelemetry-auto-instrumentation
+  ext,instrumentation: pip install {toxinidir}/opentelemetry-instrumentation
 
-  example-app: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-requests {toxinidir}/ext/opentelemetry-ext-wsgi {toxinidir}/ext/opentelemetry-ext-flask {toxinidir}/docs/examples/opentelemetry-example-app
+  example-app: pip install {toxinidir}/opentelemetry-instrumentation {toxinidir}/ext/opentelemetry-ext-requests {toxinidir}/ext/opentelemetry-ext-wsgi {toxinidir}/ext/opentelemetry-ext-flask {toxinidir}/docs/examples/opentelemetry-example-app
 
-  getting-started: pip install -e {toxinidir}/opentelemetry-auto-instrumentation -e {toxinidir}/ext/opentelemetry-ext-requests -e {toxinidir}/ext/opentelemetry-ext-wsgi -e {toxinidir}/ext/opentelemetry-ext-flask
+  getting-started: pip install -e {toxinidir}/opentelemetry-instrumentation -e {toxinidir}/ext/opentelemetry-ext-requests -e {toxinidir}/ext/opentelemetry-ext-wsgi -e {toxinidir}/ext/opentelemetry-ext-flask
 
   grpc: pip install {toxinidir}/ext/opentelemetry-ext-grpc[test]
 
   wsgi,flask,django,asgi: pip install {toxinidir}/tests/util
   wsgi,flask,django: pip install {toxinidir}/ext/opentelemetry-ext-wsgi
-  flask,django: pip install {toxinidir}/opentelemetry-auto-instrumentation
+
   asgi: pip install {toxinidir}/ext/opentelemetry-ext-asgi
 
-  boto: pip install {toxinidir}/opentelemetry-auto-instrumentation
   boto: pip install {toxinidir}/ext/opentelemetry-ext-boto[test]
 
   flask: pip install {toxinidir}/ext/opentelemetry-ext-flask[test]
@@ -204,25 +203,25 @@ commands_pre =
 
   django: pip install {toxinidir}/ext/opentelemetry-ext-django[test]
 
-  mysql: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-mysql[test]
+  mysql: pip install {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-mysql[test]
 
   opencensusexporter: pip install {toxinidir}/ext/opentelemetry-ext-opencensusexporter
 
   prometheus: pip install {toxinidir}/ext/opentelemetry-ext-prometheus
 
-  pymongo: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-pymongo[test]
+  pymongo: pip install {toxinidir}/ext/opentelemetry-ext-pymongo[test]
 
-  psycopg2: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-psycopg2 {toxinidir}/ext/opentelemetry-ext-psycopg2[test]
+  psycopg2: pip install {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-psycopg2 {toxinidir}/ext/opentelemetry-ext-psycopg2[test]
 
-  pymysql: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-pymysql[test]
+  pymysql: pip install {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-pymysql[test]
 
-  sqlite3: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-sqlite3[test]
+  sqlite3: pip install {toxinidir}/ext/opentelemetry-ext-dbapi {toxinidir}/ext/opentelemetry-ext-sqlite3[test]
 
-  redis: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-redis[test]
+  redis: pip install {toxinidir}/ext/opentelemetry-ext-redis[test]
 
-  requests: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-requests[test]
+  requests: pip install {toxinidir}/ext/opentelemetry-ext-requests[test]
 
-  jinja2: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-jinja2[test]
+  jinja2: pip install {toxinidir}/ext/opentelemetry-ext-jinja2[test]
 
   aiohttp-client: pip install {toxinidir}/opentelemetry-sdk {toxinidir}/ext/opentelemetry-ext-aiohttp-client
 
@@ -235,9 +234,8 @@ commands_pre =
   
   zipkin: pip install {toxinidir}/ext/opentelemetry-ext-zipkin
 
-  sqlalchemy: pip install {toxinidir}/opentelemetry-auto-instrumentation {toxinidir}/ext/opentelemetry-ext-sqlalchemy
+  sqlalchemy: pip install {toxinidir}/ext/opentelemetry-ext-sqlalchemy
 
-  system-metrics: pip install {toxinidir}/opentelemetry-auto-instrumentation
   system-metrics: pip install {toxinidir}/ext/opentelemetry-ext-system-metrics[test]
 
 ; In order to get a healthy coverage report,
@@ -304,9 +302,8 @@ deps =
 
 commands_pre =
   pip install -e {toxinidir}/opentelemetry-api \
-              -e {toxinidir}/opentelemetry-auto-instrumentation \
+              -e {toxinidir}/opentelemetry-instrumentation \
               -e {toxinidir}/opentelemetry-sdk \
-              -e {toxinidir}/opentelemetry-auto-instrumentation \
               -e {toxinidir}/ext/opentelemetry-ext-requests \
               -e {toxinidir}/ext/opentelemetry-ext-wsgi \
               -e {toxinidir}/ext/opentelemetry-ext-flask
@@ -331,7 +328,7 @@ changedir =
 commands_pre =
   pip install -e {toxinidir}/opentelemetry-api \
               -e {toxinidir}/opentelemetry-sdk \
-              -e {toxinidir}/opentelemetry-auto-instrumentation \
+              -e {toxinidir}/opentelemetry-instrumentation \
               -e {toxinidir}/tests/util \
               -e {toxinidir}/ext/opentelemetry-ext-dbapi \
               -e {toxinidir}/ext/opentelemetry-ext-mysql \
